--- tmeet-frontend/src/pages/meetingMake/meetingFourth.jsx
@@ -1,13 +1,0 @@
-import * as React from 'react';
-import Typography from '@mui/material/Typography';
-
-export default function Review() {
-    return (
-        <React.Fragment>
-            <Typography variant="h6" gutterBottom>
-                모임 장소 추천
-            </Typography>
-
-        </React.Fragment>
-    );
-}
