+++ tmeet-frontend/src/pages/meetingMake/place/SubwayRecommend.tsx
@@ -1,10 +1,6 @@
 import * as React from 'react';
-import NavBar from "../components/navigationBar/NavBar";
-import Container from "@mui/material/Container";
-import CssBaseline from "@mui/material/CssBaseline";
 import axios from "axios";
 import {useEffect, useState} from "react";
-import SearchList from "./SearchList.tsx"
 
 const SubwayMain = () => {
 
@@ -30,8 +26,8 @@ const SubwayMain = () => {
         const find = async () => {
             try{
                 let body = [{"subway_name": "대림역", "name": "Tom"},
-                        {"subway_name": "신정네거리", "name": "Alice"},
-                        {"subway_name": "고속터미널역", "name": "Alice"}
+                    {"subway_name": "신정네거리", "name": "Alice"},
+                    {"subway_name": "고속터미널역", "name": "Alice"}
                 ] //실제로는 사용자 입력 값 -> 세션 스토리지에 저장된 데이터 이용해서 할 것
                 const findResult = await axios.post('/api/subway', body);
                 setLat(findResult.data.res_lat)
@@ -81,21 +77,12 @@ const SubwayMain = () => {
     }
 
     return(
-        <div style={{
-            backgroundColor: "#F8FFFF",
-            width: "100%",
-            height: "200vh",
-        }}
-        >
-            <NavBar></NavBar>
+        <div>
             <h2 style={{marginTop:"50px",textAlign:"center"}}><b>모임 중간 지하철역</b> 추천 결과입니다.</h2>
             <h3 style={{marginTop:"50px",textAlign:"center", marginBottom:"10px"}}> 추천 결과 : <b>{subway_name}</b></h3>
             <h3 style={{marginTop:"10px",textAlign:"center", marginBottom:"50px"}}>예상 소요 시간 : <b>{time}</b>분</h3>
             <div style={{width:"80%", display:"block", margin:"auto"}}>
                 <div id="map" style={mapStyle}/>
-                <h3 style={{textAlign:"center", marginBottom:"30px"}}>
-                    <b>추천 모임 장소</b> 결과 리스트입니다.</h3>
-                <SearchList></SearchList>
             </div>
         </div>
     );
